+++ tests/xml/mod.rs
@@ -18760,6 +18760,26 @@ mod flex {
     }
 
     #[test]
+    fn wrap_column_auto_height_does_not_wrap_against_available_space__border_box_ltr() {
+        crate::run_xml_test("flex", "wrap_column_auto_height_does_not_wrap_against_available_space__border_box_ltr");
+    }
+
+    #[test]
+    fn wrap_column_auto_height_does_not_wrap_against_available_space__content_box_ltr() {
+        crate::run_xml_test("flex", "wrap_column_auto_height_does_not_wrap_against_available_space__content_box_ltr");
+    }
+
+    #[test]
+    fn wrap_column_auto_height_does_not_wrap_against_available_space__border_box_rtl() {
+        crate::run_xml_test("flex", "wrap_column_auto_height_does_not_wrap_against_available_space__border_box_rtl");
+    }
+
+    #[test]
+    fn wrap_column_auto_height_does_not_wrap_against_available_space__content_box_rtl() {
+        crate::run_xml_test("flex", "wrap_column_auto_height_does_not_wrap_against_available_space__content_box_rtl");
+    }
+
+    #[test]
     fn wrap_column_definite_main_size_percent_basis__border_box_ltr() {
         crate::run_xml_test("flex", "wrap_column_definite_main_size_percent_basis__border_box_ltr");
     }